+++ packages/web-components/src/components/cta/video-cta-composite.ts
@@ -9,9 +9,9 @@
 
 import { html, property, state, customElement, LitElement } from 'lit-element';
 import on from 'carbon-components/es/globals/js/misc/on.js';
-import ifNonNull from '@carbon/carbon-web-components/es/globals/directives/if-non-null.js';
-import HostListener from '@carbon/carbon-web-components/es/globals/decorators/host-listener.js';
-import HostListenerMixin from '@carbon/carbon-web-components/es/globals/mixins/host-listener.js';
+import ifNonNull from '@carbon/web-components/es/globals/directives/if-non-null.js';
+import HostListener from '@carbon/web-components/es/globals/decorators/host-listener.js';
+import HostListenerMixin from '@carbon/web-components/es/globals/mixins/host-listener.js';
 import ddsSettings from '../../internal/vendor/@carbon/ibmdotcom-utilities/utilities/settings/settings';
 import KalturaPlayerAPI from '../../internal/vendor/@carbon/ibmdotcom-services/services/KalturaPlayer/KalturaPlayer';
 import ModalRenderMixin from '../../globals/mixins/modal-render';
@@ -89,19 +89,38 @@ class DDSVideoCTAComposite extends ModalRenderMixin(
   // @ts-ignore: The decorator refers to this method but TS thinks this method is not referred to
   private async _handleRequestVideoData(event: CustomEvent) {
     const { href, videoName: customVideoName, videoDescription } = event.detail;
-    (event.target as VideoCTAMixinImpl).videoThumbnailUrl =
-      KalturaPlayerAPI.getThumbnailUrl({
-        mediaId: href,
-        width: (event?.target as HTMLElement)?.offsetWidth,
-      });
+    const videoThumbnailUrl = KalturaPlayerAPI.getThumbnailUrl({
+      mediaId: href,
+      width: (event?.target as HTMLElement)?.offsetWidth,
+    });
+    (event.target as VideoCTAMixinImpl).videoThumbnailUrl = videoThumbnailUrl;
     const videoData = await this._loadVideoData?.(href);
     if (videoData) {
       const { duration, name } = videoData;
       const videoName = customVideoName || name;
 
-      (event.target as VideoCTAMixinImpl).videoName = videoName;
-      (event.target as VideoCTAMixinImpl).videoDescription = videoDescription;
-      (event.target as VideoCTAMixinImpl).videoDuration = duration;
+      if (event.target as VideoCTAMixinImpl) {
+        (event.target as VideoCTAMixinImpl).videoName =
+          videoName || customVideoName;
+        (event.target as VideoCTAMixinImpl).videoDescription = videoDescription;
+        (event.target as VideoCTAMixinImpl).videoDuration = duration;
+      }
+
+      const videoInfo = new CustomEvent(
+        `${ddsPrefix}-cta-request-additional-video-data`,
+        {
+          bubbles: true,
+          cancelable: true,
+          composed: true,
+          detail: {
+            videoName,
+            videoDuration: duration,
+            videoThumbnailUrl,
+          },
+        }
+      );
+
+      this.dispatchEvent(videoInfo);
     }
   }
 
@@ -222,6 +241,13 @@ class DDSVideoCTAComposite extends ModalRenderMixin(
    */
   static get eventCloseLightbox() {
     return `${ddsPrefix}-expressive-modal-closed`;
+  }
+
+  /**
+   * The name of the custom event fired when there is a user gesture to run the action.
+   */
+  static get eventRequestAdditionalVideoData() {
+    return `${ddsPrefix}-cta-request-additional-video-data`;
   }
 
   /**